[ENH]: Improved error reporting (#1817)
## Description of changes

This PR reworks the way error handling works with the JavaScript client.
Previously, the api would return objects with error messages under
certain circumstances and throw errors in other circumstances. This
standardizes that by supplying a custom fetch implementation that
translates the http error codes and the server error responses, into a
standardized set of errors that are thrown by the client functions.

## Test plan

- [x] Updated javascript tests
- [x] wrote a new test 

## Documentation Changes
- This is a breaking change to the API of the client so this should be
documented.