[ENH] Integrate e2e tracing for query vectors endpoint (#1991)
## Description of changes

This PR integrates end-to-end tracing for the Query Service
(specifically the `query_vectors()` RPC). The main highlights are:
- Span inheritance by component - whenever a component is started the
executor executes the `run()` function inside a child span with the
current span as the parent. This works even across threads now for e.g.
if the component was spawned in `inherited` mode but it invoked the
receive msg handler on a different thread, etc.
- Span propagation across components. For e.g. when the HNSW Query
orchestrator submits the Pull logs request or brute force KNN request to
the worker (via the dispatcher), the worker invokes the task handler
inside a child span with parent span as the HNSW orchestrator. This is
implemented by adding a new field to the `Task` struct - the span id of
the parent and the worker then creates a child span with this id as the
parent.

## Test plan

- [+] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

Check
https://gist.github.com/sanketkedia/96f3b00a2037a0fb2e99c16c3a379e69 for
the exact logs for `query_vectors()` rpc

## Documentation Changes
None required

---------

Co-authored-by: skedia <skedia@netflix.com>