[CLN] Favor kubernetes port forwards in tilt over test loadbalancers (#2082)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
- This favors kubernetes port forwards in the tiltfile over kubernetes
services. These serve the same goal. Centralizing in tilt is cleaner
because tilt is the definition of the dev stack and its better to
minimize the test k8s yaml.
 - New functionality
	 - None

## Test plan
*How are these changes tested?*
Integration tests in CI
- [] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*