[ENH] Add otel support for query and compaction service (#2122)
## Description of changes
Export traces and spans in open telemetry format so that it can be
consumed by backends like Jaeger and Honeycomb. For Jaeger verified that
it works locally. Will test honeycomb once it is merged

## Test plan
All existing tests pass. Verified locally via tilt up and tilt down that
traces are exported to Jaeger

## Documentation Changes
NA

---------