[ENH] Use log position from sysdb to as the offset to pull from log (#2132)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
- Currently, we commit the log position for a collection in both sysdb
and log service. The commit to the sysdb must be successful for the
files visible to query nodes. The commit to offset is best effort and
can be behind in case of failures happening in between the commit to
sysdb and to the log service.
- The log position committed in the log service is used to by compactor
to figure out what which collections has new data that needs to be
compacted the offsets for the new logs. Since the log position in the
log service can be lagged and we need to use the log position in sysdb
when fetching from the log during compaction.
- This PR compares the log position in sysdb and log and use the larger
one of the 2. An alternative solution would be filtering out the log
records with offset smaller than the log position in sysdb. However,
that solution needs to inspect every log record fetched from the log
service.
 - New functionality.
 
	 - ...

## Test plan
*How are these changes tested?*

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*