[ENH] Unbox storage (#2201)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
- Unbox storage. Storage is a String and an Arc'ed type in the case of
the s3 storage impl. This does not need to be Box'ed. This was
previously implemented this way for dynamic dispatch (ala Box<dyn
Storage>) to match Box<dyn Log> and Box<dyn Sysdb>. However we now favor
enums over dyn for these cases. We should go change those sites as well.
- The string in the storage struct will be copied. We may want to
introduce https://github.com/xfbs/imstr or similar in the future for
these sorts of cases. However for this case, it is overkill.
 - New functionality
	 - None

## Test plan
*How are these changes tested?*
Existing tests. This is a non-functional change.
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*