[TST] fix enabled test_collection_with_tenant_database_overwrite test (#2293)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
- This fixes a bug in test_collection_with_tenant_database_overwrite
that was causing the test to fail. This change allows overrides of the
TenantDatabaseCollectionStateMachine to overwrite the tenant/database on
calls to create_tenant and create_database. This is a leaky abstraction,
in that its narrow and only used by one overrider. But given this is
just composition for testing, its fine.
- The singleton client will always override the tenant/database since it
should not allow creating new tenants/databases if we are using the
singleton client.
 - New functionality
	 - None

## Test plan
*How are these changes tested?*
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
None