[CLN] Make collection api object wrap model. (#2230)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
- This is cleanup borrowed from #1491. It makes Collection.py wrap a
model instead of duplicating data into collection.py and implictly
treating it as a serializable model. This is a cleaner seperation of
concerns. We have the model object, and an API wrapper used to
manipulate it via the api.
 - New functionality
	 - None

## Test plan
*How are these changes tested?*

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*