[BUG] Lock numpy to < 2.0.0 across all reqs and pyproject.toml (#2360)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
- Numpy 2.0 has some type changes which break our code. This was flagged
by @itayB in #2352 (Thank you!). We need to extend that PR a bit to
include other requirements for the thin client, as well as the
pyproject.toml.
- Note: We could fix these type issues but Onnxruntime does not yet
support python 2.0 - which we depend on
(https://github.com/microsoft/onnxruntime/issues/21063)
 - New functionality
	 - None

## Test plan
*How are these changes tested?*
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
We may want to update our docs? I am not sure what makes sense here -
input solicited.