[ENH] Materialize before bf knn (#2347)
## Description of changes
This makes the bf knn operator use materialized logs. This makes the
operator valid for cases like an invalid add etc, since those are
filtered out of the log.
- We update the output type of bf knn to not depend on logs, this is
suboptimal and highlights how the materialized log abstraction is lossy.
We should rethink this long term, and perhaps present less of a log
interface and more of a query-able layer. The lifetimes also make this
clunky, as we can't easily ship materialized log records around between
operators. For now, it follows the pattern of creating the materializer
and independently materializing logs for its operation. This is both
clunky and inefficient, but is functionally correct.
- For now, the bf operator returns the columns needed by downstream,
which has minimal impact on the orchestration and doesn't add any new
data cloning, since before we cloned off the log for returning.


Tasks
~~This is a WIP (BUT Is now ready for review minus tests, will add those
shortly)~~
- [x] Materialize before we send to log, ~~ideally in separate
operator~~
- [x] Check inputs
- [x] Refactor consumers
- [x] Rewrite tests to use Mat Log