[ENH] 1965 Split up embedding functions (#2034)
## Description of changes
- It addresses this issue:
https://github.com/chroma-core/chroma/issues/1965

There are a number of things going on in this PR. I've split the changes
in meaningful commits to facilitate the review.
- The first commit creates the empty module and renames the old
`embedding_functions.py` ffc5e9135e1e5cc471849a8cf4fc7607d287a4ff
- Then, there's one commit per embedding function moved so the reviewer
can check side by side that the contents were moved word by word.
- Above was made skipping the linters to avoid noise, so after them
there's a commit that lints the files which should be pretty innocuous .
6ad759829df125d17e6b26d20ebb987b70861705
- However, the linting over the onnx embedding function felt quite
sensible to me, so I decided to put it in a separate commit
8f08d60ebdfbe8aa7f6ef37c03a6ca1ffceb6a0e

Besides, there are a few docstrings and some tests that I'm working on
as discussed
[here](https://github.com/chroma-core/chroma/issues/1965#issuecomment-2053577220)
for which I will open a follow up PR to avoid noise here.
	 
@atroyn, can you please take a look?

## Test plan
*How are these changes tested?*
I launched a couple of times the whole test suite finding that they took
a lot of time to complete (in some 1h I had only covered 35%). The
second time I launched them, the computer even died, but not really sure
if it was because of these tests only or because something else
happening at that time.

I didn't know that there were tests for JS or rust, I will run them next
and tick the checkboxes as appropriate. I've put the outcomes in [this
comment](https://github.com/chroma-core/chroma/pull/2034#issuecomment-2068165215)

- [x] Tests for this feature pass locally with `pytest` for python,
- [ ] The whole test suite pass locally.
- [ ] `yarn test` for js, 
- [ ] `cargo test` for rust

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*

In principle all the docstrings are as they were, and users'
implementations won't be affected. However, as said before, I'm working
on a follow up PR that will add a few more tests and edit docstrings
where appropriate.

---------

Co-authored-by: atroyn <anton.troynikov@gmail.com>