[ENH] Fix issues + add sync point to test embeddings (#2397)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
- Fixes allowed_ids and disallowed_ids to also take care of
updates/deletes/upserts. For e.g. if there is an update on the log that
does not update the embedding and it is in the query list then today we
are never going to return this record even if it is in the top k
    - Adds sync points to test_embeddings + increase test timeout
    - Adds another rule in test_embeddings for compaction
    - Suppresses health check warning for filtering too much
- Fixes the case when trying to commit and flush an empty block (can
happen due to deletes). Sparse index start key can also get changed to
something that is not SparseIndexDelimiter::Start. We decided to go
ahead with flushing a dummy block if blockfile becomes fully empty so
that our segment abstraction is only uninitialized until the first
compaction; post that it is always initialized albeit with empty block
- Fixes a bug in FTS delete document where we were incorrectly panicing
- Fixes a bug in record segment apply materialization where for deletes
and updates we missed writing the max offset id
- Updates to metadata segment were missing updating the document and
were only updating the metadata
- Don't return error from the metadata segment if document is supplied
as None for an update

## Test plan
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
None