[CLN] ServerAPI to use Collection model (#2300)
## Description of changes

This PR refactors the `ServerAPI` classes so that their methods only return the `Collection` model, not the `Collection` object itself, i.e. the view on the data.

This is a much cleaner separation of concerns than we had before, and should enable faster iteration in parts. For example, the ServerAPI no longer needs to know about EmbeddingFunction or DataLoader etc - these only exist for Clients. 

We will probably re-add them when we start moving some of that functionality server-side for cloud, but for now this is a fine solution. 

Most the LoC in this PR are changing tests to use the ClientAPI and fixture instead of the ServerAPI and fixture. This does put the client in the line, but I think this is fine. 

## Test plan

Tests pass in CI. 

## Documentation Changes
N/A