[ENH] Collection Configuration Storage (#2338)
## Description of changes

This PR introduces storing collection parameters as a `Configuration`, a JSON-serializable defined object with specified configuration fields, adapting @HammadB's work in https://github.com/chroma-core/chroma/pull/1491.

Configurations are created at creation time for collections, and are immutable thereafter. 

The JSON-serialized config is stored as a new text column `config_json_str` on the `collections` table. 

Note that in this stack, Chroma does not actually consume the configuration for anything - this will be done in a separate stack. 

This PR also updates our versions of black and mypy, which were failing with our new definitions but work with the new versions. 

## Test plan

Tests pass in CI **except** the following:

- Tests involving GRPC / Cluster. This is implemented later in this stack.
- Tests involving cross-version persist. This is implemented later in this stack, along with a user-enabled migration. 
- ~~JavaScript - @jeffchuber is taking a look here.~~ Thanks @jeffchuber 

## Documentation Changes

N/A - This is not a user-facing change. Documentation around migrations are later in the stack along with the cross-version persistence tests.