[BUG] Fix 3.12 cross version persist tests by building the wheel manu… (#2482)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
- Now that we build hnswlib wheels for 3.12, the old version of chroma
fetches wheels which do not work. This change explicitly makes the cross
version tests rebuild the wheels. This was silently failing as the
incompatible wheels failed in python with invalid instruction. Testing
if the deps work out of the box for old versions is out of scope for
this test. That should be handled by the existing tests that promote a
given version, not retroactive tests.
 - New functionality
	 - None

## Test plan
*How are these changes tested?*
These fix tests on main
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*