[ENH] Synchronize post s3.get() operations in providers (#2712)
*Summarize the changes made by this PR.*
Adds synchronization points in hnsw_provider and blockfile provider for
concurrent cache inserts to the same key. This prevents dangling
references to entries that the cache is not aware of. For the HNSW
provider, we also synchronize writing out the index byte buffer to the
local files since that is not guaranteed to be thread safe.

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

None