[ENH] - Fix `get_or_create_collection` semantic inconsistencies (#2810)
## Description of changes
`get_or_create` collection was overriding metadata on the "get" path
when the collection already exists. This PR removes this behavior to
make the API consistent with its name. Users can utilize
`Collection.modify` to change a collection's metadata.

One idea for `get_or_create` was to rename arguments prefix arguments
with `create_`, to make it clear that they only apply to the `create`
path. We decided to remove the overriding behavior on the `get` path
without renaming the arguments to minimize breakage.

Another idea for future consideration is to add a
`Client.modify_collection` API. This would make it clear at a quick
glance that a separate modifying API is available. A caveat is that we
already have a `Collection.modify` API.

## Test plan
- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
Documentation updated to reflect this change.

Co-authored-by: Itai Smith <itaismith@Itais-MacBook-Pro.local>