[BUG] Pull logs should fail if requested offset has been purged (#2746)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
- It can happen that log purging purges a set of records after which a
request is made to pull these logs. This can happen after a compaction
if there was an inflight query that got the offset from sysdb before the
compaction. This PR fails such a request. The query will fail as a
result and the client is welcome to retry.

## Test plan
*How are these changes tested?*
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
None