[PERF] Convert embeddings representation to numpy (#2803)
## Description of changes

*Summarize the changes made by this PR.*
**Improvements & Bug fixes**
- Standardizes our representation of embedding vectors to numpy arrays
instead of python lists

**Notes**

- The conversion of a numpy array to a python list via `.tolist()`
consumes an excess amount of memory, and so the focus of this PR is to
reduce the number of these operations by representing embeddings as
numpy arrays across the board.
- The main concern is consuming too much memory on our users' machines.
thus the focus is to reduce meory usage within Local Chroma. This
naturally will necessitate changes to Distributed, the approach is to
only make minimal, necessary changes to the latter.

- The format for `Embeddings` changes from `List[List[Union[float,
int]]]` to `List[NDArray[Union[int32, float32]]]`.
- Using a 2D numpy array instead of a Python list of 1D numpy arrays
would be slightly more performant. However, it would require deep,
fundamental, and widespread changes to not only our code but also
testing frameworks. I have a commit that starts to go down this path (if
you're interested in seeing), and it's not pretty. The changes introduce
an incredible amount of risk across the board for little to no gain on
the vector we are trying to improve and thus is not worth it.
   
- We tested that when creating lists of numpy arrays, Python stores
pointers to the arrays and doesn't consume excess memory—using a 2D
array thus doesn't help us further achieve our goal of reduced memory
usage.
- FastAPI does not support serialization of numpy arrays. Thus, upon
encoding/decoding embeddings through our middleware, we have no option
but to convert back to python lists. This is acceptable, however,
because these operations only happen within Distributed. Issue to fix
here: https://github.com/chroma-core/chroma/issues/2844
- This PR introduces a breaking change to the API in that embeddings
will now be returned as 2D numpy arrays. It makes sense to me both from
a memory usage issue perspective (not calling `.tolist()` upon returning
a response) and also generally provides better performance within users.
- Embeddings will now be internally standardized to `float32`/`int32`
upon serialization (for RPC and byte encoding), which will could cause a
little bit of noise within distance functions depending on what the user
provides. We should communicate this change in the next version release.
- Because of the aforementioned FastAPI serialization issue and the fact
that structures like `QueryResult` and `GetResult` are threaded all the
way across the stack through our middleware, there is a new risk
introduced. We could fork Local/Distributed Chroma and create new
structures for the former that use numpy arrays. However, this is
unideal to me. I decided instead to let these structures support both
numpy arrays and pythonic lists for embeddings. When were are on either
side of the middleware, we must convert between numpy arrays and lists.
This isn’t an issue in of itself, but makes it such that we always have
to remember to perform the conversion when using the structures, lest we
risk returning embeddings as pythonic lists within Distributed Chroma.
New development of clients seems unlikely for now, so this feels like an
acceptable risk.
- Using the same profiling method, over 10 runs the average memory used
was originally 362 MB, going to 264 MB with changes (representing a 28%
decrease in memory usage).

<details>
<summary>Performance details</summary>
I ran the following code 10 times each on main and on my branch and
averaged the results to test the performance. This was the same code as
used in the original
[issue](https://github.com/chroma-core/chroma/issues/2665)

```
import chromadb
import numpy as np

NUM_RECORDS = 4096 * 4 

ids = [str(i) for i in range(NUM_RECORDS)]
embeddings = np.random.rand(NUM_RECORDS, 384)
metadatas = [{"name": f"metadata_{i}"} for i in range(NUM_RECORDS)]

client = chromadb.Client()
collection = client.create_collection("test_collection")

collection.add(ids, embeddings)
```
</details>

## Test plan
*How are these changes tested?*

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust