[ENH] Implement serde for SparseIndex (#2881)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
	 - Implements the serde traits for the sparse index
- Refactors out the forward/reverse mapping into one lock protected
struct we serde from
	 - Clean up imports on the benchmark crate, since that uses bincode.
 - New functionality
	 - None

Notes
- I chose to implement this with the default serde impls because then we
get zero-copy into the in memory struct. Since the serde is used for
on-disk caching, that is preferable. A potential alternative would be to
serialize to/from bytes and implement our own Serializer/Deserializer.
- I forwarded the pattern of exposing the "forward" of the SparseIndex.
This is a bad abstraction and I'll clean it up in a subsequent PR,
preferring to keep it as is for now.

## Test plan
*How are these changes tested?*
Added a basic unit test using
[bincode](https://github.com/bincode-org/bincode)
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
None