[ENH]  Make block and sparse index caches use persistent type.
End-to-end tests confirm that with this change foyer will initialize the
disk path (assuming it's writable) and instantiate a disk.  A memory
cache will also instantiate.

Note that the PersistentCache type means that the key and value
implement StorageKey and StorageValue, so it's totally acceptable for
every non-persistent cache to implement the trait.  This means that any
place that takes a persistent cache will also take the non-persistent
versions.  It's trait logic, not "is a".