[CLN] Make delete return None (#2880)
## Description of changes

The `Collection.delete` API seemingly returns None, but our internal
implementation returns a list of the IDs that were deleted. This PR
makes all `delete` implementations return `None` for consistency.

## Test plan
*How are these changes tested?*

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
The documentation does not mention the API returning a list of deleted
IDs, so it is consistent with the expected behavior.