[ENH] Fix error propagation in sysdb (#2896)
## Description of changes

*Summarize the changes made by this PR.*
- RPC errors in sysdb were set in the response protobufs of sysdb in the
"status" field. Upstream consumers in a few places don't check this
status field at all and thus ignore errors from sysdb. In one case, this
led to the log service deleting collections because it incorrectly
assumed a collection to be not present whereas in reality the rpc had
failed. This PR deprecates this field from the protobuf and instead
transforms the error to a grpc error that grpc sends to the consumers
separately
- Also, added better logging messages in the sysdb. It's a piece of
trash rn
- Please note that I have deliberately not made this backwards
compatible. The only issues that can arise would be during rolling
upgrade deployment for the first version that has this change and given
our scale, I am comfortable with dealing with it. Since this change is
very fundamental to how we propagate errors in grpc instead of
maintaining this field and its backward compatibility for the
foreseeable future, the sooner we do it the right way the better

## Test plan
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
None