[CLN] Separate validation and transformation logic (#2899)
## Description of changes

We perform validation, and transformations, on requests before we send
them from the client to the server. Previously, these operations were
not well modularized, leading to difficult to debug code, lots of
exceptional paths, and overall not-niceness.

This PR separates validation, and makes it idemponent, from
transformations, for all methods on `Collection` in the python client.

Along the way, this PR cleans up the possibility of sending empty lists
or dicts as filters or lists of IDs, so that the tests are now
consistent with our validators.

In a future PR, we will also repeat this same validation logic
server-side for use with clients other than our first party, and update
it in JavaScript.

## Test plan

CI Passes

New validation to ensure empty filters trigger validators. 

## Documentation Changes

## TODO
 - [ ] Technically this is a breaking API change, let's discuss