[ENH] Sometimes simulation disagrees with math (#3030)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
- One can view the assignment as a binomial distribution. We can take
the std dev of the binomial to validate the null hypothesis - that our
data is a binomial distribution.
- As @Sicheng-Pan pointed out
(https://github.com/chroma-core/chroma/pull/2969#discussion_r1821694613)
the previous tolerance to the multi-assign test should have failed. It
did not. Sometimes math disagrees with simulation. We were still in two
sigma with a tolerance of 75, now the tolerance is 150.
 - New functionality
	 - None

## Test plan
*How are these changes tested?*
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*