[ENH]: refine some default arguments for foyer (#3037)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
	 - Refine some default arguments for the foyer configuraiton.
 - New functionality
	 - ...

Please note that the foyer manages the lifecycle of disk data through
regions (in the file system, each region is a file). The lifecycle of a
region will dynamically change based on workload, the number of
`flushers`, the number of `reclaimers`, and the size of
`clean_region_threshold`. In the configuration, ensure that the number
of regions is much greater than the total number of flushers and
reclaimers (clean_region_threshold is generally between 1 to 1.5 times
the number of reclaimers). At the same time, the number of reclaimers
should be between 0.5 to 1 times the number of flushers.

In the file system, the calculation of the number of regions is
`capacity / file size`.

In the production environment, the `capacity` size and
`admission_rate_limit` should be increased. The `capacity` size should
be the size of the disk cache you need, and the `admission_rate_limit`
should be the write bandwidth reserved for the disk cache (determined by
the bandwidth of EBS).

## Test plan
*How are these changes tested?*

- [ ] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*

Signed-off-by: MrCroxx <mrcroxx@outlook.com>