[BUG] Fix max seq ID incorrect bug by reducing state (#3064)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
- There was a bug where if the ID already existed the max seq id of the
batch was not advanced. This results in storing more data than we need
to on the log since the seq id is underreported. Not a huge issue, but
breaks tests. This change removes `max_seq_id` from the batch entirely.
`local_hnsw` will track it correctly in memory, and
`local_persistent_hnsw` tracks it correctly on persist advances. For
example, `add A` and then `add A` will result in the second add A to
note advance the `batch` max_seq_id, by removing the needless redundant
state we prevent the error.
 - New functionality
	 - None

## Test plan
*How are these changes tested?*
This is a test fix. I've included the hypothesis repro as a new test. I
ran test_persist.py 15 times locally with cleared hypothesis
generations.
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
None