[ENH]  Evicting object store.
This object store tracks the size of the objects inserted to it and
deletes objects when the object store's size exceeds a configured
threshold.  Tests show that it will prune on startup as appropriate.

Before we ship this with a configuration language (a missing step that
enables us to use this code), I'd like to investigate ways to prevent an
evicting object storage from accidentally evicting 100% of an S3 bucket.