[ENH] Implement `KnnOrchestrator` (#2998)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
	 - N/A
 - New functionality
- Implement everything relevant for the new `KnnFilterOrchestrator` and
`KnnOrchestrator`. They are seperated because a query shares the same
filter but has different embeddings to query for. We simply spin up a
single `KnnFilterOrchestrator` and then spin up multiple
`KnnOrchestrator` in parallel, instead of tracking the states in a
single orchestrator for different embeddings.
- The `KnnFilterOrchestrator` only generates allowed/exclude offset ids.
- The `KnnOrchestrator` computes the KNN given a single query embedding.
- The `KnnOrchestrator` runs KNN on the log with `KnnLogOperator` and on
the hnsw with `KnnHnswOperator`. The result is merged by
`KnnMergeOperator` and hydrated by `KNNProjectOperator`.

## Test plan
*How are these changes tested?*

- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
*Are all docstrings for user-facing APIs updated if required? Do we need
to make documentation changes in the [docs
repository](https://github.com/chroma-core/docs)?*

---------

Co-authored-by: Sicheng Pan <sicheng@trychroma.com>