[CHORE] Make block cache weight 8. Make cache config unit based. (#3115)
## Description of changes

*Summarize the changes made by this PR.*
 - Improvements & Bug fixes
- The cache code uses weighted capacity and specifies the capacity in
MiB units. But we currently assume blocks are 1 MB.
- The underlying foyer cache is capacity and weight based. Fixing to MiB
makes cache configuration indirect from the mechanics of the cache and
results in it being cumbersome and potentially outright impossible to
use. For example if I want to limit to 7000 items. I have to specify a
very small value as my capacity. This value may not be an integer but
the configuration expects an integer capacity. Or alternatively, I can
say that each item has some byte weight and choose a limit based on
that. Rather than force these workarounds, it makes more sense for
callers to think in whatever units they need and make the cache
specification agnostic to that. As is, if a cache does not want to think
in MiB and instead wants to think in items, it cannot without
complication.
 - New functionality
	 - None

## Test plan
*How are these changes tested?*
- [x] Tests pass locally with `pytest` for python, `yarn test` for js,
`cargo test` for rust

## Documentation Changes
None